Fix ratio error

+++ src/main/kotlin/xyz/atrius/waystones/service/WorldRatioService.kt
@@ -3,8 +3,8 @@ package xyz.atrius.waystones.service
 import org.bukkit.World
 import xyz.atrius.waystones.data.JsonFile
 
-object WorldRatioService : JsonFile<Int>("ratios") {
+object WorldRatioService : JsonFile<Double>("ratios") {
 
-    operator fun get(world: World): Int =
-        data.getOrDefault(world.name, 1)
+    operator fun get(world: World): Double =
+        data.getOrDefault(world.name, 1.0)
 }
