+++ src/main/kotlin/xyz/atrius/waystones/utility/Location.kt
@@ -87,5 +87,5 @@ fun Location.sameDimension(world: World?) =
 fun Location.playSound(sound: Sound, volume: Float = 1f, pitch: Float = 1f) = Bukkit.getOnlinePlayers()
     .forEach { if (it.world == world) it.playSound(this, sound, volume, pitch)  }
 
-fun World.getRatio(): Int =
+fun World.getRatio(): Double =
     WorldRatioService[this]
